[style] 데이터 타입 변경 (#2)
데이터 타입 변경

+++ app/src/main/java/com/example/githubapi/model/BasicResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class BasicResponse(
     @SerializedName("total_count")
-    val totalCount: Int,
+    val totalCount: Long,
     @SerializedName("incomplete_results")
     val incompleteResults: Boolean,
     val items: List<ItemsData>
